+++ resources/deployable.rb
@@ -14,7 +14,7 @@
 # limitations under the License.
 #
 
-actions :deploy, :undeploy
+actions :deploy, :undeploy, :enable, :disable
 
 attribute :component_name, :kind_of => String, :name_attribute => true
 attribute :version, :kind_of => String, :required => true